sdl3::audio
Trait AudioFormatNum

pub trait AudioFormatNum: Copy + 'static {
    const SILENCE: Self;

    // Required method
    fn audio_format() -> AudioFormat;
}

A phantom type for retrieving the SDL_AudioFormat of a given generic type. All format types are returned as native-endian.

Required Associated Constants

const SILENCE: Self

The appropriately typed silence value for the audio format used.

Examples

// The AudioFormatNum trait has to be imported for the Channel::SILENCE part to work.
use sdl3::audio::{AudioCallback, AudioFormatNum};

struct Silence;

impl<Channel> AudioCallback<Channel> for Silence
where
    Channel: AudioFormatNum,
{
    fn callback(&mut self, out: &mut [Channel]) {
        for dst in out.iter_mut() {
            *dst = Channel::SILENCE;
       }
    }
}

Required Methods

fn audio_format() -> AudioFormat

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.