sdl3
Module filesystem

Structs

EnumerationResult

Possible results from an enumeration callback.

Folder

The type of the OS-provided default folder for a specific purpose.

GlobFlags

GlobResults

GlobResultsIter

PathInfo

PathType

Types of filesystem entries.

Enums

FileSystemError

PrefPathError

Functions

copy_file

create_directory

enumerate_directory

get_base_path

get_path_info

get_pref_path

Return the preferred directory for the application to write files on this system, based on the given organization and application name.

get_user_folder

glob_directory

remove_path

rename_path

Type Aliases

EnumerateCallback