sdl3::filesystem
Struct Folder

#[repr(transparent)]
pub struct Folder(pub i32);

The type of the OS-provided default folder for a specific purpose.

Note that the Trash folder isn’t included here, because trashing files usually involves extra OS-specific functionality to remember the file’s original location.

The folders supported per platform are:

Windows	macOS/iOS	tvOS	Unix (XDG)	Haiku	Emscripten

HOME	X	X		X	X	X

DESKTOP	X	X		X	X

DOCUMENTS	X	X		X

DOWNLOADS	Vista+	X		X

MUSIC	X	X		X

PICTURES	X	X		X

PUBLICSHARE		X		X

SAVEDGAMES	Vista+

SCREENSHOTS	Vista+

TEMPLATES	X	X		X

VIDEOS	X	X*		X

Note that on macOS/iOS, the Videos folder is called “Movies”.

Availability

This enum is available since SDL 3.2.0.

See also

SDL_GetUserFolder

Known values (sdl3-sys)

Associated constant	Global constant	Description

HOME	SDL_FOLDER_HOME	The folder which contains all of the current user’s data, preferences, and documents. It usually contains most of the other folders. If a requested folder does not exist, the home folder can be considered a safe fallback to store a user’s documents.

DESKTOP	SDL_FOLDER_DESKTOP	The folder of files that are displayed on the desktop. Note that the existence of a desktop folder does not guarantee that the system does show icons on its desktop; certain GNU/Linux distros with a graphical environment may not have desktop icons.

DOCUMENTS	SDL_FOLDER_DOCUMENTS	User document files, possibly application-specific. This is a good place to save a user’s projects.

DOWNLOADS	SDL_FOLDER_DOWNLOADS	Standard folder for user files downloaded from the internet.

MUSIC	SDL_FOLDER_MUSIC	Music files that can be played using a standard music player (mp3, ogg…).

PICTURES	SDL_FOLDER_PICTURES	Image files that can be displayed using a standard viewer (png, jpg…).

PUBLICSHARE	SDL_FOLDER_PUBLICSHARE	Files that are meant to be shared with other users on the same computer.

SAVEDGAMES	SDL_FOLDER_SAVEDGAMES	Save files for games.

SCREENSHOTS	SDL_FOLDER_SCREENSHOTS	Application screenshots.

TEMPLATES	SDL_FOLDER_TEMPLATES	Template files to be used when the user requests the desktop environment to create a new file in a certain folder, such as “New Text File.txt”. Any file in the Templates folder can be used as a starting point for a new file.

VIDEOS	SDL_FOLDER_VIDEOS	Video files that can be played using a standard video player (mp4, webm…).

COUNT	SDL_FOLDER_COUNT	Total number of types in this enum, not a folder type by itself.

Tuple Fields
0: i32
Implementations

impl SDL_Folder

pub const HOME: SDL_Folder

The folder which contains all of the current user’s data, preferences, and documents. It usually contains most of the other folders. If a requested folder does not exist, the home folder can be considered a safe fallback to store a user’s documents.

pub const DESKTOP: SDL_Folder

The folder of files that are displayed on the desktop. Note that the existence of a desktop folder does not guarantee that the system does show icons on its desktop; certain GNU/Linux distros with a graphical environment may not have desktop icons.

pub const DOCUMENTS: SDL_Folder

User document files, possibly application-specific. This is a good place to save a user’s projects.

pub const DOWNLOADS: SDL_Folder

Standard folder for user files downloaded from the internet.

pub const MUSIC: SDL_Folder

Music files that can be played using a standard music player (mp3, ogg…).

pub const PICTURES: SDL_Folder

Image files that can be displayed using a standard viewer (png, jpg…).

pub const PUBLICSHARE: SDL_Folder

Files that are meant to be shared with other users on the same computer.

pub const SAVEDGAMES: SDL_Folder

Save files for games.

pub const SCREENSHOTS: SDL_Folder

Application screenshots.

pub const TEMPLATES: SDL_Folder

Template files to be used when the user requests the desktop environment to create a new file in a certain folder, such as “New Text File.txt”. Any file in the Templates folder can be used as a starting point for a new file.

pub const VIDEOS: SDL_Folder

Video files that can be played using a standard video player (mp4, webm…).

pub const COUNT: SDL_Folder

Total number of types in this enum, not a folder type by itself.