sdl3::gamepad
Struct Gamepad

pub struct Gamepad { /* private fields */ }

Wrapper around the SDL_Gamepad object

Implementations

impl Gamepad

pub fn subsystem(&self) -> &GamepadSubsystem

pub fn name(&self) -> String

Return the name of the controller or an empty string if no name is found.

pub fn mapping(&self) -> String

Return a String describing the controller’s button and axis mappings

pub fn attached(&self) -> bool

Return true if the controller has been opened and currently connected.

pub fn instance_id(&self) -> u32

Return the joystick instance id of this controller

pub fn axis(&self, axis: Axis) -> i16

Get the position of the given axis

pub fn button(&self, button: Button) -> bool

Returns true if button is pressed.

pub fn set_rumble( &mut self, low_frequency_rumble: u16, high_frequency_rumble: u16, duration_ms: u32, ) -> Result<(), IntegerOrSdlError>

Set the rumble motors to their specified intensities, if supported. Automatically resets back to zero after duration_ms milliseconds have passed.

Notes

The value range for the intensities is 0 to 0xFFFF.

Do not use std::u32::MAX or similar for duration_ms if you want the rumble effect to keep playing for a long time, as this results in the effect ending immediately after starting due to an overflow. Use some smaller, “huge enough” number instead.

pub fn set_rumble_triggers( &mut self, left_rumble: u16, right_rumble: u16, duration_ms: u32, ) -> Result<(), IntegerOrSdlError>

Start a rumble effect in the game controller’s triggers.

pub unsafe fn has_led(&self) -> bool

Query whether a game controller has a RGB LED.

pub unsafe fn has_rumble(&self) -> bool

Query whether a game controller has rumble support.

pub unsafe fn has_rumble_triggers(&self) -> bool

Query whether a game controller has rumble support on triggers.

pub fn set_led( &mut self, red: u8, green: u8, blue: u8, ) -> Result<(), IntegerOrSdlError>

Update a game controller’s LED color.

pub fn send_effect(&mut self, data: &[u8]) -> Result<(), Error>

Send a controller specific effect packet.