sdl3::gpu
Struct SamplerCreateInfo

pub struct SamplerCreateInfo { /* private fields */ }

Implementations

impl SamplerCreateInfo

pub fn new() -> Self

pub fn with_min_filter(self, filter: Filter) -> Self

The minification filter to apply to lookups.

pub fn with_mag_filter(self, filter: Filter) -> Self

The magnification filter to apply to lookups.

pub fn with_mipmap_mode(self, mode: SamplerMipmapMode) -> Self

The mipmap filter to apply to lookups.

pub fn with_address_mode_u(self, mode: SamplerAddressMode) -> Self

The addressing mode for U coordinates outside [0, 1).

pub fn with_address_mode_v(self, mode: SamplerAddressMode) -> Self

The addressing mode for V coordinates outside [0, 1).

pub fn with_address_mode_w(self, mode: SamplerAddressMode) -> Self

The addressing mode for W coordinates outside [0, 1).

pub fn with_mip_lod_bias(self, value: f32) -> Self

The bias to be added to mipmap LOD calculation.

pub fn with_max_anisotropy(self, value: f32) -> Self

The anisotropy value clamp used by the sampler. If enable_anisotropy is false, this is ignored.

pub fn with_compare_op(self, value: CompareOp) -> Self

The comparison operator to apply to fetched data before filtering.

pub fn with_min_lod(self, value: f32) -> Self

Clamps the minimum of the computed LOD value.

pub fn with_max_lod(self, value: f32) -> Self

Clamps the maximum of the computed LOD value.

pub fn with_enable_anisotropy(self, enable: bool) -> Self

True to enable anisotropic filtering.

pub fn with_enable_compare(self, enable: bool) -> Self

True to enable comparison against a reference value during lookups.