sdl3::gpu
Struct TextureCreateInfo

pub struct TextureCreateInfo { /* private fields */ }

Implementations

impl TextureCreateInfo

pub fn new() -> Self

pub fn with_type(self, value: TextureType) -> Self

The base dimensionality of the texture.

pub fn with_format(self, format: TextureFormat) -> Self

The pixel format of the texture.

pub fn with_usage(self, value: TextureUsage) -> Self

How the texture is intended to be used by the client.

pub fn with_width(self, value: u32) -> Self

The width of the texture.

pub fn with_height(self, value: u32) -> Self

The height of the texture.

pub fn with_layer_count_or_depth(self, value: u32) -> Self

The layer count or depth of the texture. This value is treated as a layer count on 2D array textures, and as a depth value on 3D textures.

pub fn with_num_levels(self, value: u32) -> Self

The number of mip levels in the texture.

pub fn with_sample_count(self, value: SampleCount) -> Self

The number of samples per texel. Only applies if the texture is used as a render target.