sdl3::iostream
Struct IOStream

pub struct IOStream<'a> { /* private fields */ }

A structure that provides an abstract interface to stream I/O.

Implementations

impl<'a> IOStream<'a>

pub unsafe fn raw(&self) -> *mut SDL_IOStream

pub unsafe fn from_ll<'b>(raw: *mut SDL_IOStream) -> IOStream<'b>

pub fn from_file<P: AsRef<Path>>( path: P, mode: &str, ) -> Result<IOStream<'static>, Error>

Creates an SDL file stream.

pub fn from_bytes(buf: &'a [u8]) -> Result<IOStream<'a>, Error>

Prepares a read-only memory buffer for use with IOStream.

This method can only fail if the buffer size is zero.

pub fn from_read<T>( r: &mut T, buffer: &'a mut Vec<u8>, ) -> Result<IOStream<'a>, Error>
where T: Read + Sized,

Reads a Read object into a buffer and then passes it to IOStream.from_bytes.

The buffer must be provided to this function and must live as long as the IOStream, but the IOStream does not take ownership of it.

pub fn from_bytes_mut(buf: &'a mut [u8]) -> Result<IOStream<'a>, Error>

Prepares a read-write memory buffer for use with IOStream.

This method can only fail if the buffer size is zero.

pub fn len(&self) -> Option<usize>

Gets the streamโs total size in bytes.

Returns None if the stream size canโt be determined (either because it doesnโt make sense for the stream type, or there was an error).

pub fn is_empty(&self) -> bool