sdl3::joystick
Struct Joystick

pub struct Joystick { /* private fields */ }

Wrapper around the SDL_Joystick object

Implementations

impl Joystick

pub const fn subsystem(&self) -> &JoystickSubsystem

pub fn name(&self) -> String

Return the name of the joystick or an empty string if no name is found.

pub fn attached(&self) -> bool

Return true if the joystick has been opened and currently connected.

pub fn instance_id(&self) -> u32

pub fn guid(&self) -> Guid

Retrieve the joystick’s GUID

pub fn power_level(&self) -> Result<PowerInfo, IntegerOrSdlError>

Retrieve the battery level of this joystick

pub fn num_axes(&self) -> u32

Retrieve the number of axes for this joystick

pub fn axis(&self, axis: u32) -> Result<i16, IntegerOrSdlError>

Gets the position of the given axis.

The function will fail if the joystick doesn’t have the provided axis.

pub fn num_buttons(&self) -> u32

Retrieve the number of buttons for this joystick

pub fn button(&self, button: u32) -> Result<bool, IntegerOrSdlError>

Return Ok(true) if button is pressed.

The function will fail if the joystick doesn’t have the provided button.

pub fn num_hats(&self) -> u32

Retrieve the number of balls for this joystick

pub fn hat(&self, hat: u32) -> Result<HatState, IntegerOrSdlError>

Return the position of hat for this joystick

pub fn set_rumble( &mut self, low_frequency_rumble: u16, high_frequency_rumble: u16, duration_ms: u32, ) -> bool

Set the rumble motors to their specified intensities, if supported. Automatically resets back to zero after duration_ms milliseconds have passed.

Notes

The value range for the intensities is 0 to 0xFFFF.

Do not use std::u32::MAX or similar for duration_ms if you want the rumble effect to keep playing for a long time, as this results in the effect ending immediately after starting due to an overflow. Use some smaller, “huge enough” number instead.

Returns false if the gamepad doesn’t support rumble.

pub fn set_rumble_triggers( &mut self, left_rumble: u16, right_rumble: u16, duration_ms: u32, ) -> Result<(), IntegerOrSdlError>

Start a rumble effect in the joystick’s triggers.

pub unsafe fn has_led(&self) -> bool

Query whether a joystick has an LED.

pub unsafe fn has_rumble(&self) -> bool

Query whether a joystick has rumble support.

pub unsafe fn has_rumble_triggers(&self) -> bool

Query whether a joystick has rumble support on triggers.

pub fn set_led( &mut self, red: u8, green: u8, blue: u8, ) -> Result<(), IntegerOrSdlError>

Update a joystick’s LED color.

pub fn send_effect(&mut self, data: &[u8]) -> Result<(), IntegerOrSdlError>

Send a joystick specific effect packet.