sdl3::messagebox
Function show_simple_message_box

pub fn show_simple_message_box<'a, W>(
    flags: MessageBoxFlag,
    title: &str,
    message: &str,
    window: W,
) -> Result<(), ShowMessageError>
where
    W: Into<Option<&'a Window>>,

Show a simple message box, meant to be informative only.

There is no way to know if the user clicked “Ok” or closed the message box, If you want to retrieve which button was clicked and customize a bit more your message box, use show_message_box instead.