sdl3
Struct VideoSubsystem

pub struct VideoSubsystem { /* private fields */ }

Implementations

impl VideoSubsystem

pub fn clipboard(&self) -> ClipboardUtil

impl VideoSubsystem

pub fn text_input(&self) -> TextInputUtil

impl VideoSubsystem

pub fn sdl(&self) -> Sdl

Obtain an SDL context.

impl VideoSubsystem

pub fn gl_attr(&self) -> GLAttr<'_>

Obtains access to the OpenGL window attributes.

impl VideoSubsystem

pub fn window(&self, title: &str, width: u32, height: u32) -> WindowBuilder

Initializes a new WindowBuilder; a convenience method that calls WindowBuilder::new().

pub fn window_and_renderer( &self, title: &str, width: u32, height: u32, ) -> Result<WindowCanvas, Error>

Create a window with a renderer.

pub unsafe fn popup_window( &self, window: &Window, width: u32, height: u32, ) -> PopupWindowBuilder

Initializes a new PopupWindowBuilder; a convenience method that calls PopupWindowBuilder::new().

pub fn current_video_driver(&self) -> &'static str

pub fn num_video_drivers(&self) -> Result<i32, Error>

pub fn displays(&self) -> Result<Vec<Display>, Error>

pub fn get_primary_display(&self) -> Result<Display, Error>

Get primary display ID.

pub fn is_screen_saver_enabled(&self) -> bool

pub fn enable_screen_saver(&self)

pub fn disable_screen_saver(&self)

pub fn gl_load_library_default(&self) -> Result<(), Error>

Loads the default OpenGL library.

This should be done after initializing the video driver, but before creating any OpenGL windows. If no OpenGL library is loaded, the default library will be loaded upon creation of the first OpenGL window.

If a different library is already loaded, this function will return an error.

pub fn gl_load_library<P: AsRef<Path>>(&self, path: P) -> Result<(), Error>

Loads the OpenGL library using a platform-dependent OpenGL library name (usually a file path).

This should be done after initializing the video driver, but before creating any OpenGL windows. If no OpenGL library is loaded, the default library will be loaded upon creation of the first OpenGL window.

If a different library is already loaded, this function will return an error.

pub fn gl_unload_library(&self)

Unloads the current OpenGL library.

To completely unload the library, this should be called for every successful load of the OpenGL library.

pub fn gl_get_proc_address(&self, procname: &str) -> SDL_FunctionPointer

Gets the pointer to the named OpenGL function.

This is useful for OpenGL wrappers such as gl-rs.

pub fn gl_extension_supported(&self, extension: &str) -> bool

pub fn gl_get_current_window_id(&self) -> Result<u32, Error>

pub fn gl_release_current_context(&self) -> Result<(), Error>

Releases the threadโs current OpenGL context, i.e. sets the current OpenGL context to nothing.

pub fn gl_set_swap_interval<S: Into<SwapInterval>>( &self, interval: S, ) -> Result<(), Error>

pub fn gl_get_swap_interval(&self) -> Result<SwapInterval, Error>

pub fn vulkan_load_library_default(&self) -> Result<(), Error>

Loads the default Vulkan library.

This should be done after initializing the video driver, but before creating any Vulkan windows. If no Vulkan library is loaded, the default library will be loaded upon creation of the first Vulkan window.

If a different library is already loaded, this function will return an error.

pub fn vulkan_load_library<P: AsRef<Path>>(&self, path: P) -> Result<(), Error>

Loads the Vulkan library using a platform-dependent Vulkan library name (usually a file path).

This should be done after initializing the video driver, but before creating any Vulkan windows. If no Vulkan library is loaded, the default library will be loaded upon creation of the first Vulkan window.

If a different library is already loaded, this function will return an error.

pub fn vulkan_unload_library(&self)

Unloads the current Vulkan library.

To completely unload the library, this should be called for every successful load of the Vulkan library.

pub fn vulkan_get_proc_address_function(&self) -> SDL_FunctionPointer

Gets the pointer to the vkGetInstanceProcAddr Vulkan function. This function can be called to retrieve the address of other Vulkan functions.

pub fn get_system_theme() -> SystemTheme

Get the current system theme.