sdl3::surface
Struct SurfaceRef

pub struct SurfaceRef { /* private fields */ }

An unsized Surface reference.

This type is used whenever Surfaces need to be borrowed from the SDL library, without concern for freeing the Surface.

Implementations

impl SurfaceRef

pub unsafe fn from_ll<'a>(raw: *const SDL_Surface) -> &'a SurfaceRef

pub unsafe fn from_ll_mut<'a>(raw: *mut SDL_Surface) -> &'a mut SurfaceRef

pub fn raw(&self) -> *mut SDL_Surface

pub fn width(&self) -> u32

pub fn height(&self) -> u32

pub fn pitch(&self) -> u32

pub fn size(&self) -> (u32, u32)

pub fn rect(&self) -> Rect

Gets the rect of the surface.

pub fn pixel_format(&self) -> PixelFormat

pub fn pixel_format_enum(&self) -> PixelFormat

pub fn with_lock<R, F: FnOnce(&[u8]) -> R>(&self, f: F) -> R

Locks a surface so that the pixels can be directly accessed safely.

pub fn with_lock_mut<R, F: FnOnce(&mut [u8]) -> R>(&mut self, f: F) -> R

Locks a surface so that the pixels can be directly accessed safely.

pub unsafe fn without_lock(&self) -> Option<&[u8]>

Returns the Surface’s pixel buffer if the Surface doesn’t require locking (e.g. it’s a software surface).

pub unsafe fn without_lock_mut(&mut self) -> Option<&mut [u8]>

Returns the Surface’s pixel buffer if the Surface doesn’t require locking (e.g. it’s a software surface).

pub unsafe fn must_lock(&self) -> bool

Returns true if the Surface needs to be locked before accessing the Surface pixels.

pub fn save_bmp_rw(&self, iostream: &mut IOStream<'_>) -> Result<(), Error>

pub fn save_bmp<P: AsRef<Path>>(&self, path: P) -> Result<(), Error>

pub fn set_palette(&mut self, palette: &Palette) -> Result<(), Error>

pub fn enable_RLE(&mut self)

pub fn disable_RLE(&mut self)

pub fn set_color_key(&mut self, enable: bool, color: Color) -> Result<(), Error>

pub fn color_key(&self) -> Result<Color, Error>

The function will fail if the surface doesn’t have color key enabled.

pub fn set_color_mod(&mut self, color: Color)

pub fn color_mod(&self) -> Color

pub fn fill_rect<R>(&mut self, rect: R, color: Color) -> Result<(), Error>
where R: Into<Option<Rect>>,

pub fn fill_rects(&mut self, rects: &[Rect], color: Color) -> Result<(), Error>

pub fn set_alpha_mod(&mut self, alpha: u8)

pub fn alpha_mod(&self) -> u8

pub fn set_blend_mode(&mut self, mode: BlendMode) -> Result<(), Error>

The function will fail if the blend mode is not supported by SDL.

pub fn blend_mode(&self) -> BlendMode

pub fn set_clip_rect<R>(&mut self, rect: R) -> bool
where R: Into<Option<Rect>>,

Sets the clip rectangle for the surface.

If the rectangle is None, clipping will be disabled.

pub fn clip_rect(&self) -> Option<Rect>

Gets the clip rectangle for the surface.

Returns None if clipping is disabled.

pub fn convert(&self, format: &PixelFormat) -> Result<Surface<'static>, Error>

Copies the surface into a new one that is optimized for blitting to a surface of a specified pixel format.

pub fn convert_format( &self, format: PixelFormat, ) -> Result<Surface<'static>, Error>

Copies the surface into a new one of a specified pixel format.

pub fn blit<R1, R2>( &self, src_rect: R1, dst: &mut SurfaceRef, dst_rect: R2, ) -> Result<Option<Rect>, Error>
where R1: Into<Option<Rect>>, R2: Into<Option<Rect>>,

Performs surface blitting (surface copying).

Returns the final blit rectangle, if a dst_rect was provided.

pub unsafe fn lower_blit<R1, R2>( &self, src_rect: R1, dst: &mut SurfaceRef, dst_rect: R2, ) -> Result<(), Error>
where R1: Into<Option<Rect>>, R2: Into<Option<Rect>>,

Performs low-level surface blitting.

Unless you know what you’re doing, use blit() instead, which will clip the input rectangles. This function could crash if the rectangles aren’t pre-clipped to the surface, and is therefore unsafe.

pub unsafe fn soft_stretch_linear<R1, R2>( &self, src_rect: R1, dst: &mut SurfaceRef, dst_rect: R2, ) -> Result<Option<Rect>, Error>
where R1: Into<Option<Rect>>, R2: Into<Option<Rect>>,

Performs bilinear scaling between two surfaces of the same format, 32BPP.

Returns the final blit rectangle, if a dst_rect was provided.

pub fn blit_scaled<R1, R2>( &self, src_rect: R1, dst: &mut SurfaceRef, dst_rect: R2, scale_mode: SDL_ScaleMode, ) -> Result<Option<Rect>, Error>
where R1: Into<Option<Rect>>, R2: Into<Option<Rect>>,

Performs scaled surface bliting (surface copying).

Returns the final blit rectangle, if a dst_rect was provided.

pub unsafe fn lower_blit_scaled<R1, R2>( &self, src_rect: R1, dst: &mut SurfaceRef, dst_rect: R2, scale_mode: SDL_ScaleMode, ) -> Result<(), Error>
where R1: Into<Option<Rect>>, R2: Into<Option<Rect>>,

Performs low-level scaled surface blitting.

Unless you know what you’re doing, use blit_scaled() instead, which will clip the input rectangles. This function could crash if the rectangles aren’t pre-clipped to the surface, and is therefore unsafe.