sdl3
Module video

Modules

gl_attr

OpenGL context getters and setters

Structs

Display

DisplayMode

DriverIterator

GLContext

PopupWindowBuilder

The type that allows you to build popup windows.

Window

Represents the “shell” of a Window.

WindowBuilder

The type that allows you to build windows.

WindowContext

Holds a SDL_Window

WindowSurfaceRef

Enums

FlashOperation

Represents a setting for a window flash operation.

FullscreenType

GLProfile

Orientation

Represents orientation of a display.

SwapInterval

Represents a setting for vsync/swap interval.

SystemTheme

System theme.

WindowBuildError

WindowPos

Functions

drivers

Gets an iterator of all video drivers compiled into the SDL2 library.

Type Aliases

VkInstance

(sdl3-sys) Enable a use-ash-* feature to alias this to vk::Instance from the ash crate. Otherwise it’s a pointer to an opaque struct.

VkSurfaceKHR