sdl3::video
Struct PopupWindowBuilder

pub struct PopupWindowBuilder { /* private fields */ }

The type that allows you to build popup windows.

Implementations

impl PopupWindowBuilder

pub unsafe fn new( v: &VideoSubsystem, parent_window: &Window, width: u32, height: u32, ) -> PopupWindowBuilder

Initializes a new PopupWindowBuilder.

pub fn build(&self) -> Result<Window, WindowBuildError>

Builds the popup window

pub fn window_flags(&self) -> u32

Gets the underlying window flags.

pub fn set_window_flags(&mut self, flags: u32) -> &mut PopupWindowBuilder

Sets the underlying window flags. This will effectively undo any previous build operations, excluding window size and position.

pub fn offset(&mut self, x: i32, y: i32) -> &mut PopupWindowBuilder

Sets the window offset relative to the parent window.

pub fn opengl(&mut self) -> &mut PopupWindowBuilder

Sets the window to be usable with an OpenGL context

pub fn vulkan(&mut self) -> &mut PopupWindowBuilder

Sets the window to be usable with a Vulkan instance

pub fn hidden(&mut self) -> &mut PopupWindowBuilder

Hides the window.

pub fn resizable(&mut self) -> &mut PopupWindowBuilder

Sets the window to be resizable.

pub fn input_grabbed(&mut self) -> &mut PopupWindowBuilder

Sets the window to have grabbed input focus.

pub fn metal_view(&mut self) -> &mut PopupWindowBuilder

Create a SDL_MetalView when constructing the window. This is required when using the raw_window_handle feature on MacOS. Has no effect no other platforms.

pub fn tooltip(&mut self) -> &mut PopupWindowBuilder

Sets the window to be a tooltip.

pub fn popup_menu(&mut self) -> &mut PopupWindowBuilder

Sets the window to be a popup menu.

pub fn transparent(&mut self) -> &mut PopupWindowBuilder

Sets the window to be transparent

pub fn always_on_top(&mut self) -> &mut PopupWindowBuilder

Sets the window to be shown on top of all other windows